Phase 1: Backend MVP — tested and working
- Fix TypeScript build errors (missing axios dep, BigInt types, DTO fields)
- Fix NestJS dependency injection (AuthModule exports JwtModule)
- Remap Docker ports (Postgres:5434, Redis:6381) to avoid conflicts
- Add gateway module for WebSocket chat gateway
- Add database seed script with test tenant, admin, channels, players
- Add 40 unit tests for chat, moderation, tenant services and API key guard
- Update CLAUDE.md with actual Phase 1 status

+++ packages/backend/src/modules/webhook/webhook.service.ts
@@ -139,8 +139,8 @@ export class WebhookService {
         tenantId,
         channelId,
         initiatedById: 'webhook',
-        totalAmount: BigInt(Math.floor(amount * 100)) / 100n,
-        perPlayerAmount: BigInt(Math.floor(perPlayerAmount * 100)) / 100n,
+        totalAmount: Math.floor(amount * 100) / 100,
+        perPlayerAmount: Math.floor(perPlayerAmount * 100) / 100,
         durationSeconds,
       },
     });
